requirements.txt
Django>=5.0,<6.0
pytest>=7.0.0
pytest-django>=4.5.2
factory-boy>=3.2.1
